+++ tests/test_analyse.py
@@ -139,9 +139,9 @@ def test_sector_ratio():
 def test_basic_stats():
     bw.basic_stats(DATA)
     bs2 = bw.basic_stats(DATA['Spd80mN'])
-    assert (bs2['count'][0] == 95180.0) and (round(bs2['mean'][0], 6) == 7.518636) and \
-           (round(bs2['std'][0], 6) == 3.994552) and (round(bs2['min'][0], 3) == 0.215) and \
-           (round(bs2['max'][0], 1) == 29.0)
+    assert (bs2['count'].iloc[0] == 95180.0) and (round(bs2['mean'].iloc[0], 6) == 7.518636) and \
+           (round(bs2['std'].iloc[0], 6) == 3.994552) and (round(bs2['min'].iloc[0], 3) == 0.215) and \
+           (round(bs2['max'].iloc[0], 1) == 29.0)
 
 
 def test_time_continuity_gaps():
@@ -209,9 +209,9 @@ def test_ti_twelve_by_24():
 
 def test_coverage():
     # hourly coverage
-    assert round(bw.coverage(DATA[['Spd80mN']], period='1H')[
+    assert round(bw.coverage(DATA[['Spd80mN']], period='1h')[
            '2016-01-09 17:00':'2016-01-09 17:30'].values[0][0], 5) == 0.83333
-    assert round(bw.coverage(DATA.Spd80mN, period='1H')[
+    assert round(bw.coverage(DATA.Spd80mN, period='1h')[
            '2016-01-09 17:00':'2016-01-09 17:30'].values[0], 5) == 0.83333
     # monthly_coverage
     assert round(bw.coverage(DATA.Spd80mN, period='1M')['2016-05-01'], 5) == 0.36537