+++ tests/test_correlation.py
@@ -102,7 +102,7 @@ def test_ordinary_least_squares():
     assert round(correl.params['num_data_points'], 5) == correl_monthly_results_90_intercept['num_data_points']
 
     # check hourly, checked against Excel
-    correl = bw.Correl.OrdinaryLeastSquares(MERRA2_NE['WS50m_m/s'], DATA_CLND['Spd80mN'], averaging_prd='1H',
+    correl = bw.Correl.OrdinaryLeastSquares(MERRA2_NE['WS50m_m/s'], DATA_CLND['Spd80mN'], averaging_prd='1h',
                                             coverage_threshold=1)
     correl.run()
     assert round(correl.params['slope'], 5) == correl_hourly_results['slope']
@@ -113,7 +113,7 @@ def test_ordinary_least_squares():
     # check aggregation method
     correl_aggregation_results = {'slope': 5.98789, 'offset': -9.32585, 'r2': 0.9304, 'num_data_points': 12445}
     correl = bw.Correl.OrdinaryLeastSquares(MERRA2_NE['T2M_degC'], DATA_CLND['T2m'],
-                                            averaging_prd='1H', coverage_threshold=1,
+                                            averaging_prd='1h', coverage_threshold=1,
                                             ref_aggregation_method='sum', target_aggregation_method='sum')
     correl.run()
     correl.plot()
@@ -277,7 +277,7 @@ def test_synthesize():
 
     for idx, row in pd.DataFrame(result_ord_lst_sq).iterrows():
         # Comparing the first 6 digits to avoid issuing with floating point precision
-        assert str(row[0])[0:6] == str(synth.loc[idx][0])[0:6]
+        assert str(row.iloc[0])[0:6] == str(synth.loc[idx].iloc[0])[0:6]
 
     # Test the synthesise for when the ref_dir is given as input.
     correl = bw.Correl.OrdinaryLeastSquares(MERRA2_NE['WS50m_m/s']['2016-03-02 00:00:00':],
@@ -288,7 +288,7 @@ def test_synthesize():
     synth = correl.synthesize()
 
     for idx, row in pd.DataFrame(result_ord_lst_sq_dir).iterrows():
-        assert str(row[0]) == str(round(synth.loc[idx][0], 6))
+        assert str(row.iloc[0]) == str(round(synth.loc[idx].iloc[0], 6))
 
     # Test the synthesise when SpeedSort correlation is used.
     correl = bw.Correl.SpeedSort(MERRA2_NE['WS50m_m/s']['2016-03-02 00:00:00':'2017-03-02 00:00:00'],
@@ -301,14 +301,15 @@ def test_synthesize():
 
     for idx, row in pd.DataFrame(result_speed_sort).iterrows():
         print(idx)
-        assert str(row[0]) == str(round(synth.loc[idx][0], 6))
+        assert str(row.iloc[0]) == str(round(synth.loc[idx].iloc[0], 6))
 
     # Test the synthesise when SpeedSort correlation is used using 10 min averaging period.
     data_test = DATA_CLND[['Spd80mN', 'Spd60mN', 'Dir78mS', 'Dir58mS']].copy()
-    data_test['Dir78mS']['2016-01-09 17:10:00':'2016-01-09 17:50:00'] = np.nan
-    data_test['Spd80mN']['2016-01-09 17:10:00':'2016-01-09 17:50:00'] = np.nan
-    data_test['Dir58mS']['2016-01-09 17:50:00':'2016-01-10 19:10:00'] = np.nan
-    data_test['Spd60mN']['2016-01-09 17:50:00':'2016-01-10 19:10:00'] = np.nan
+    data_test.loc['2016-01-09 17:10:00':'2016-01-09 17:50:00', 'Dir78mS'] = np.nan
+    data_test.loc['2016-01-09 17:10:00':'2016-01-09 17:50:00', 'Spd80mN'] = np.nan
+    data_test.loc['2016-01-09 17:50:00':'2016-01-10 19:10:00', 'Dir58mS'] = np.nan
+    data_test.loc['2016-01-09 17:50:00':'2016-01-10 19:10:00', 'Spd60mN'] = np.nan
+
     ss_cor = bw.Correl.SpeedSort(data_test['Spd80mN'], data_test['Dir78mS'], data_test['Spd60mN'], data_test['Dir58mS'],
                                  averaging_prd='10min')
     ss_cor.run()
@@ -370,7 +371,7 @@ def test_orthogonal_least_squares():
     assert round(correl.params['num_data_points'], 5) == correl_monthly_results['num_data_points']
 
     # check hourly
-    correl = bw.Correl.OrthogonalLeastSquares(MERRA2_NE['WS50m_m/s'], DATA_CLND['Spd80mN'], averaging_prd='1H',
+    correl = bw.Correl.OrthogonalLeastSquares(MERRA2_NE['WS50m_m/s'], DATA_CLND['Spd80mN'], averaging_prd='1h',
                                               coverage_threshold=1)
     correl.run()
     assert round(correl.params['slope'], 5) == correl_hourly_results['slope']
@@ -381,7 +382,7 @@ def test_orthogonal_least_squares():
     # check aggregation method
     correl_aggregation_results = {'slope': 6.42434, 'offset': -12.8301, 'r2': 0.9255, 'num_data_points': 12445}
     correl = bw.Correl.OrthogonalLeastSquares(MERRA2_NE['T2M_degC'], DATA_CLND['T2m'],
-                                              averaging_prd='1H', coverage_threshold=1,
+                                              averaging_prd='1h', coverage_threshold=1,
                                               ref_aggregation_method='sum', target_aggregation_method='sum')
     correl.run()
     assert round(correl.params['slope'], 5) == correl_aggregation_results['slope']
@@ -403,7 +404,7 @@ def test_multiple_linear_regression():
     # check aggregation method
     correl_aggregation_results = {'slope': [5.51666, 0.54769], 'offset': -10.44818, 'num_data_pts': 12445}
     correl = bw.Correl.MultipleLinearRegression([MERRA2_NE['T2M_degC'], MERRA2_NW['T2M_degC']], DATA_CLND['T2m'],
-                                                averaging_prd='1H', coverage_threshold=1,
+                                                averaging_prd='1h', coverage_threshold=1,
                                                 ref_aggregation_method='sum', target_aggregation_method='sum')
     correl.run()
     for idx, slope in enumerate(correl.params['slope']):
@@ -570,7 +571,7 @@ def test_speed_sort():
     }
     ss_cor = bw.Correl.SpeedSort(MERRA2_NE['WS50m_m/s'], MERRA2_NE['WD50m_deg'],
                                  DATA_CLND['Spd80mN'], DATA_CLND['Dir78mS'],
-                                 averaging_prd='1H')
+                                 averaging_prd='1h')
     ss_cor.run()
 
     assert ss_cor.params['overall_average_veer'] == result['overall_average_veer']